+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-## Added
+### Changed
+
+- Move third party services to adapters [(#90)](https://github.com/plotwist-app/plotwist-backend/pull/90)
+
+### Added
 
 - Consume SQS messages [(#71)](https://github.com/plotwist-app/plotwist-backend/pull/71)
 - Add review summary [(#87)](https://github.com/plotwist-app/plotwist-backend/pull/87)